Avoid warning message

+++ geopdes/src/op_curlu_curlv_2d.cc
@@ -108,7 +108,8 @@ OUTPUT: \n\
           } 
         else
           {
-            warning_with_id ("geopdes:zero_measure_element", "op_curlu_curlv_2d: element %d has 0 area", iel);
+            warning_with_id ("geopdes:zero_measure_element", "op_curlu_curlv_2d: element %lld has 0 area", 
+            static_cast<long long int> (iel));
           }  // end for iel, if area > 0
       
       if (nargout == 1)
