+++ geopdes/src/op_uxn_vxn_2d.cc
@@ -110,7 +110,8 @@ OUTPUT:\n\
             } 
           else
             {
-              {warning_with_id ("geopdes:zero_measure_element", "op_uxn_vxn_2d: element %d has 0 area", iel);}
+              {warning_with_id ("geopdes:zero_measure_element", "op_uxn_vxn_2d: element %lld has 0 area", 
+            static_cast<long long int> (iel));}
             }  // end for iel, if area > 0      
       } //end of parallel section
