+++ geopdes/src/op_uxn_vxn_3d.cc
@@ -153,7 +153,8 @@ OUTPUT:\n\
                     } // end for jdof
                 } // end for idof
             } else {
-            {warning_with_id ("geopdes:zero_measure_element", "op_uxn_vxn_3d: element %d has 0 area", iel);}
+            {warning_with_id ("geopdes:zero_measure_element", "op_uxn_vxn_3d: element %lld has 0 area", 
+            static_cast<long long int> (iel));}
           }  // end for iel, if area > 0
       } // end of parallel section
